+++ Domain/Entities/Recurso.cs
@@ -13,7 +13,7 @@ namespace Domain.Entities
     {
         [Key]
         public int RecursoId { get; set; }
-        public int userid { get; set; }
+        public Int16 userid { get; set; }
         [ForeignKey("userid")]
         public virtual User User { get; set; }
         public int competenciaid { get; set; }
